requirements.txt
youtube_transcript_api
openai
python-dotenv
requests
streamlit